implement Linked List
a.	Create a linked list.
b.	Insert an element at the start of the linked list.
c.	Insert an element at the end of the linked list.
d.	Insert an element before an existing element whose information is x in a linked list.
e.	Insert an element after an existing element whose information is x in a linked list.
f.	Delete the first element of the linked list.
g.	Delete the last element of the linked list.
h.	Delete the element whose information is x from a linked list.
i.	Display the contents of the linked list.