count leaf,non-leaf,height and check binary tree
a.	Count the number of nodes in a binary tree.
b.	Count the number of leaf nodes in a binary tree.
c.	Count the number of non-leaf nodes in a binary tree.
d.	Return the height of the binary tree.
e.	Check whether the tree is a strict binary tree or not.